NF Ajax form auto-check class

+++ os-sim/www/net/modifynet.php
@@ -42,6 +42,92 @@ require_once ('classes/Net_scan.inc');
 require_once ('classes/Util.inc');
 
 Session::logcheck("MenuPolicy", "PolicyNetworks");
+
+$error = false;
+
+$net_name     = POST('netname');
+$cidr         = POST('cidr');
+$descr        = POST('descr');
+$asset        = POST('asset');
+$sensors      = ( isset($_POST['sboxs'] ) && !empty ( $_POST['sboxs']) ) ? Util::clean_array(POST('sboxs')) : array();
+$threshold_a  = POST('threshold_a');
+$threshold_c  = POST('threshold_c');
+$rrd_profile  = POST('rrd_profile');
+$nagios       = POST('nagios');
+$clone        = POST('clone');
+
+$num_sensors = count($sensors);
+
+$validate = array (
+	"netname"     => array("validation"=>"OSS_NET_NAME", "e_message" => 'illegal:' . _("Network Name")),
+	"cidr"        => array("validation"=>"OSS_IP_CIDR", "e_message" => 'illegal:' . _("CIDR")),
+	"descr"       => array("validation"=>"OSS_ALPHA, OSS_NULLABLE, OSS_SPACE, OSS_PUNC, OSS_AT, OSS_NL", "e_message" => 'illegal:' . _("Description")),
+	"asset"       => array("validation"=>"OSS_DIGIT", "e_message" => 'illegal:' . _("Asset")),
+	"sboxs"       => array("validation"=>"OSS_ALPHA, OSS_SCORE, OSS_PUNC, OSS_AT", "e_message" => 'illegal:' . _("Sensors")),
+	"rrd_profile" => array("validation"=>"OSS_ALPHA, OSS_NULLABLE, OSS_SPACE, OSS_PUNC", "e_message" => 'illegal:' . _("RRD Profile")),
+	"threshold_a" => array("validation"=>"OSS_DIGIT", "e_message" => 'illegal:' . _("Threshold A")),
+	"threshold_c" => array("validation"=>"OSS_DIGIT", "e_message" => 'illegal:' . _("Threshold C")));
+
+$clone = POST('clone'); // for Duplicate selected
+
+if ( GET('ajax_validation') == true )
+{
+	$validation_errors = validate_form_fields('GET', $validate);
+	if ( $validation_errors == 1 )
+		echo 1;
+	else if ( empty($validation_errors) )
+		echo 0;
+	else
+		echo $validation_errors[0];
+		
+	exit();
+}
+else
+{
+	$validation_errors = validate_form_fields('POST', $validate);
+	
+	if ( ( $validation_errors == 1 ) ||  (is_array($validation_errors) && !empty($validation_errors)) || $num_sensors == 0 )
+	{
+		$error = true;
+		
+		$message_error = array();
+				
+		if( $num_sensors == 0)
+			$message_error [] = _("You Need to select at least one Sensor");
+			
+		if ( is_array($validation_errors) && !empty($validation_errors) )
+			$message_error = array_merge($message_error, $validation_errors);
+		else
+		{
+			if ($validation_errors == 1)
+				$message_error [] = _("Invalid send method");
+		}	
+	}
+	
+	if ( POST('ajax_validation_all') == true )
+	{
+		if ( is_array($message_error) && !empty($message_error) )
+			echo implode( "<br/>", $message_error);
+		else
+			echo 0;
+		
+		exit();
+	}
+}
+
+if ( $error == true )
+{
+	$_SESSION['_net']['net_name']    = $net_name;
+	$_SESSION['_net']['cidr']        = $cidr;
+	$_SESSION['_net']['descr']       = $descr;
+	$_SESSION['_net']['asset']       = $asset;
+	$_SESSION['_net']['sensors']     = $sensors;
+	$_SESSION['_net']['threshold_a'] = $threshold_a;
+	$_SESSION['_net']['threshold_c'] = $threshold_c;
+	$_SESSION['_net']['rrd_profile'] = $rrd_profile;
+	$_SESSION['_net']['nagios']      = $nagios;
+}
+
 ?>
 
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
@@ -52,74 +138,53 @@ Session::logcheck("MenuPolicy", "PolicyNetworks");
 	<meta http-equiv="Pragma" content="no-cache">
 	<link type="text/css" rel="stylesheet" href="../style/style.css"/>
 </head>
+
 <body>
-                                                                                
-<h1> <?php echo gettext("Update net"); ?> </h1>
+
+<?php
+if (GET('withoutmenu') != "1") 
+	include ("../hmenu.php"); 
+?>
+
+<h1> <?php echo gettext("Update Network"); ?> </h1>   
+
 
 <?php
 
-$net_name = POST('name');
-$threshold_a = POST('threshold_a');
-$threshold_c = POST('threshold_c');
-$asset = POST('asset');
-$descr = POST('descr');
-$nsens = POST('nsens');
-$ips = POST('ips');
-$alert = POST('alert');
-$persistence = POST('persistence');
-$rrd_profile = POST('rrd_profile');
-$clone = POST('clone'); // for Duplicate selected
 
-ossim_valid($net_name, OSS_NET_NAME, 'illegal:' . _("Net name"));
-ossim_valid($ips, OSS_ALPHA, OSS_SPACE, OSS_PUNC, 'illegal:' . _("Ips"));
-ossim_valid($asset, OSS_DIGIT, 'illegal:' . _("Asset"));
-ossim_valid($threshold_a, OSS_DIGIT, 'illegal:' . _("threshold_a"));
-ossim_valid($threshold_c, OSS_DIGIT, 'illegal:' . _("threshold_c"));
-ossim_valid($nsens, OSS_DIGIT, OSS_NULLABLE, 'illegal:' . _("nnets"));
-ossim_valid($alert, OSS_DIGIT, OSS_NULLABLE, 'illegal:' . _("Alert"));
-ossim_valid($clone, OSS_DIGIT, OSS_NULLABLE, 'illegal:' . _("clone"));
-ossim_valid($persistence, OSS_DIGIT, OSS_NULLABLE, 'illegal:' . _("Persistence"));
-ossim_valid($rrd_profile, OSS_ALPHA, OSS_NULLABLE, OSS_SPACE, OSS_PUNC, 'illegal:' . _("Net name"));
-ossim_valid($descr, OSS_ALPHA, OSS_NULLABLE, OSS_SPACE, OSS_PUNC, OSS_AT, 'illegal:' . _("Description"));
-if (ossim_error()) {
-    die(ossim_error());
-}
-if (POST('insert'))
+if ( POST('insert') && !empty($net_name) )
 {
-    $sensors = array();
-    for ($i = 1; $i <= $nsens; $i++)
+    if ( $error == true)
 	{
-        $name = "mboxs" . $i;
-        ossim_valid(POST($name) , OSS_NULLABLE, OSS_ALPHA, OSS_PUNC, OSS_SPACE);
-        if (ossim_error()) {
-            die(ossim_error());
-        }
-        $name_aux = POST($name);
-        if (!empty($name_aux)) $sensors[] = POST($name);
-    }
-    if (!count($sensors)) {
-        die(ossim_error(_("At least one sensor is required")));
-    }
-    
+		$txt_error = "<div>"._("We Found the following errors").":</div><div style='padding:10px;'>".implode( "<br/>", $message_error)."</div>";				
+		Util::print_error($txt_error);	
+		Util::make_form("POST", "newnetform.php?name=".$net_name);
+		die();
+	}
+		
     $db = new ossim_db();
     $conn = $db->connect();
+		
     if ($clone)
-		Net::insert($conn, $net_name, $ips, $asset, $threshold_c, $threshold_a, $rrd_profile, $alert, $persistence, $sensors, $descr);
+		Net::insert($conn, $net_name, $cidr, $asset, $threshold_c, $threshold_a, $rrd_profile, $alert, $persistence, $sensors, $descr);
 	else 
-		Net::update($conn, $net_name, $ips, $asset, $threshold_c, $threshold_a, $rrd_profile, $alert, $persistence, $sensors, $descr);
+		Net::update($conn, $net_name, $cidr, $asset, $threshold_c, $threshold_a, $rrd_profile, $alert, $persistence, $sensors, $descr);
     
 	Net_scan::delete($conn, $net_name, 3001);
     
 	Net_scan::delete($conn, $net_name, 2007);
     
-	if (POST('nessus')) 
-        Net_scan::insert($conn, $net_name, 3001, 0);
+	//if (POST('nessus')) Net_scan::insert($conn, $net_name, 3001, 0);
     
-    if (POST('nagios'))
+    if ( $nagios )
         Net_scan::insert($conn, $net_name, 2007, 0);
     
     $db->close($conn);
 }
+
+if ( isset($_SESSION['_net']) )
+	unset($_SESSION['_net']);
+	
 ?>
     <p> <?php echo gettext("Net succesfully updated"); ?> </p>
     <script>document.location.href="net.php"</script>
